Added My-Tickets page and fixed some bugs

+++ project.php
@@ -13,12 +13,15 @@ $project = $projects->ProjectDetails($_GET["proj_id"]);
 //show only OPEN and IN_PROGRESS tickets
 $statuses = array('status' => "'OPEN', 'IN_PROGRESS'");
 $allTickets = $tickets->GetAllProjectsTickets($project->id, $statuses['status']);
+//show CLOSED tickets
+$closedStatus = array('status' => "'CLOSED'");
+$allClosedTickets = $tickets->GetAllProjectsTickets($project->id, $closedStatus['status']);
 //
 $usersOnProject = $projects->GetAssignedUsers($project->id);
 $unassignedUsers = $projects->GetUnassignedUsers($project->id);
 
 if ((!$users->UserIsAdmin($user->id)) && (!$projects->CheckUserAssignedOnProject($user->id, $project->id))) {
-    $msg_errors = "You are not allowed to view this project!";
+    $msg_errors = "You are not allowed to access that URL!";
     $_SESSION["msg_error"] = $msg_errors;
     header("location:my-projects.php");
     exit;
@@ -26,6 +29,7 @@ if ((!$users->UserIsAdmin($user->id)) && (!$projects->CheckUserAssignedOnProject
 
 //Start ticket
 if (isset($_GET["start_ticket"]) && !empty($_GET["start_ticket"])) {
+    $ticket = $tickets->TicketDetails($_GET["start_ticket"]);
     if (($users->UserIsAdmin($user->id) !== true) && ($user->id !== $ticket->user_id)) {
         $msg_errors = "You are not allowed to start Ticket #" . $_GET["start_ticket"];
         $_SESSION["proj_msg_error"] = $msg_errors;
@@ -48,16 +52,24 @@ if (isset($_GET["start_ticket"]) && !empty($_GET["start_ticket"])) {
 
 //Delete ticket
 if (isset($_GET["del_ticket"]) && !empty($_GET["del_ticket"])) {
-    if ($tickets->DeleteTicket($_GET["del_ticket"])) {
-        $msg_success = "Ticket deleted successfully!";
-        $_SESSION["proj_msg_success"] = $msg_success;
-        header("location:project.php?proj_id=" . $project->id);
-        exit;
-    } else {
-        $msg_errors = "Couldn't delete the ticket!";
+    $ticket = $tickets->TicketDetails($_GET["del_ticket"]);
+    if (($users->UserIsAdmin($user->id) !== true) && ($user->id !== $ticket->user_id)) {
+        $msg_errors = "You are not allowed to delete Ticket #" . $_GET["del_ticket"];
         $_SESSION["proj_msg_error"] = $msg_errors;
         header("location:project.php?proj_id=" . $project->id);
         exit;
+    } else {
+        if ($tickets->DeleteTicket($_GET["del_ticket"])) {
+            $msg_success = "Ticket deleted successfully!";
+            $_SESSION["proj_msg_success"] = $msg_success;
+            header("location:project.php?proj_id=" . $project->id);
+            exit;
+        } else {
+            $msg_errors = "Couldn't delete the ticket!";
+            $_SESSION["proj_msg_error"] = $msg_errors;
+            header("location:project.php?proj_id=" . $project->id);
+            exit;
+        }
     }
 }
 
@@ -185,6 +197,7 @@ include "includes/header.php";
                                         <tr>
                                             <th scope="col">Actions</th>
                                             <th scope="col">#</th>
+                                            <th scope="col">Project</th>
                                             <th scope="col">Title</th>
                                             <th scope="col">Status</th>
                                             <th scope="col">User</th>
@@ -211,17 +224,18 @@ include "includes/header.php";
                                                 <a href="edit-ticket.php?id=<?php echo $ticket->id; ?>"
                                                     title="Edit Ticket" class="ticket-actions"><i
                                                         class="far fa-edit edit-ticket"></i></a>
-                                                <?php if ($users->UserIsAdmin($user->id)): ?><a
+                                                <?php if ($users->UserIsAdmin($user->id) || $user->id == $ticket->user_id): ?><a
                                                     href="project.php?proj_id=<?php echo $ticket->project_id; ?>&del_ticket=<?php echo $ticket->id; ?>"
                                                     title="Delete Ticket" class="ticket-actions"><i
                                                         class="far fa-trash-alt delete-ticket"></i></a><?php endif;?>
                                             </td>
                                             <td><?php echo $ticket->id; ?></td>
+                                            <td><?php echo $project->proj_name; ?></td>
                                             <td><?php echo $ticket->title; ?></td>
                                             <td><?php echo $ticket->status; ?></td>
                                             <td><?php echo $ticket->user_fullname; ?></td>
                                             <td><?php echo $ticket->create_date; ?></td>
-                                            <td><?php echo ($ticket->start_date !== $ticket->create_date) ? $ticket->start_date : ""; ?>
+                                            <td><?php echo $ticket->start_date; ?>
                                             </td>
                                         </tr>
                                         <?php endforeach;?>
@@ -233,7 +247,7 @@ include "includes/header.php";
                 </div>
             </div>
 
-            <div class="row add-project mt-0">
+            <div class="row mt-0">
                 <div class="col-12 m-auto">
                     <?php if (!empty($usersOnProject)): ?>
                     <div class="card card-assigned-users">
@@ -325,9 +339,6 @@ include "includes/header.php";
                                         <button type="submit" class="btn btn-secondary btn-block btn-delete"
                                             name="btnDelete">DELETE PROJECT</button>
                                     </div>
-                                    <div class="copyright">2020 &copy; <a href="https://www.ButcoSoft.com"
-                                            class="copy-link">ButcoSoft</a>. All
-                                        rights reserved.</div>
                                 </div>
                             </form>
                         </div>
@@ -335,6 +346,62 @@ include "includes/header.php";
                 </div>
                 <?php endif;?>
             </div>
+            <div class="row closed-tickets">
+                <div class="col-12 m-auto">
+                    <div class="card">
+                        <div class="card-title text-center mt-3 mb-0">
+                            <h4>CLOSED Tickets</h4>
+                        </div>
+                        <div class="card-body">
+                            <div class="table-responsive-xl">
+                                <table class="table table-hover table-sm">
+                                    <thead>
+                                        <tr>
+                                            <th scope="col">Actions</th>
+                                            <th scope="col">#</th>
+                                            <th scope="col">Project</th>
+                                            <th scope="col">Title</th>
+                                            <th scope="col">Status</th>
+                                            <th scope="col">User</th>
+                                            <th scope="col">Started At</th>
+                                            <th scope="col">Closed At</th>
+                                        </tr>
+                                    </thead>
+                                    <tbody>
+                                        <?php foreach ($allClosedTickets as $ticket): ?>
+                                        <tr>
+                                            <td>
+                                                <a href="ticket.php?id=<?php echo $ticket->id; ?>" title="View Ticket"
+                                                    class="ticket-actions"><i class="far fa-eye view-ticket"></i></a>
+                                                <a href="edit-ticket.php?id=<?php echo $ticket->id; ?>"
+                                                    title="Edit Ticket" class="ticket-actions"><i
+                                                        class="far fa-edit edit-ticket"></i></a>
+                                                <?php if ($users->UserIsAdmin($user->id) || $user->id == $ticket->user_id): ?><a
+                                                    href="project.php?proj_id=<?php echo $ticket->project_id; ?>&del_ticket=<?php echo $ticket->id; ?>"
+                                                    title="Delete Ticket" class="ticket-actions"><i
+                                                        class="far fa-trash-alt delete-ticket"></i></a><?php endif;?>
+                                            </td>
+                                            <td><?php echo $ticket->id; ?></td>
+                                            <td><?php echo $project->proj_name; ?></td>
+                                            <td><?php echo $ticket->title; ?></td>
+                                            <td><?php echo $ticket->status; ?></td>
+                                            <td><?php echo $ticket->user_fullname; ?></td>
+                                            <td><?php echo $ticket->start_date; ?></td>
+                                            <td><?php echo $ticket->close_date; ?>
+                                            </td>
+                                        </tr>
+                                        <?php endforeach;?>
+                                    </tbody>
+                                </table>
+                                <div class="copyright">2020 &copy; <a href="https://www.ButcoSoft.com"
+                                        class="copy-link">ButcoSoft</a>. All
+                                    rights reserved.</div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+
         </div>
     </div>
 </div>
